Prefer int over int32_t

+++ include/FastNoise/Generators/Generator.inl
@@ -74,7 +74,7 @@ public:
         return simdT;
     }
 
-    FastNoise::OutputMinMax GenUniformGrid2D( float* noiseOut, int32_t xStart, int32_t yStart, int32_t xSize, int32_t ySize, float frequency, int32_t seed ) const final
+    FastNoise::OutputMinMax GenUniformGrid2D( float* noiseOut, int xStart, int yStart, int xSize, int ySize, float frequency, int seed ) const final
     {
         float32v min( INFINITY );
         float32v max( -INFINITY );
@@ -87,14 +87,14 @@ public:
         int32v xSizeV( xSize );
         int32v xMax = xSizeV + xIdx + int32v( -1 );
 
-        int32_t totalValues = xSize * ySize;
-        int32_t index = 0;
+        size_t totalValues = xSize * ySize;
+        size_t index = 0;
 
         xIdx += int32v::FS_Incremented();
 
         AxisReset<true>( xIdx, yIdx, xMax, xSizeV, xSize );
 
-        while( index < totalValues - (int32_t)FS_Size_32() )
+        while( index < totalValues - FS_Size_32() )
         {
             float32v xPos = FS_Converti32_f32( xIdx ) * freqV;
             float32v yPos = FS_Converti32_f32( yIdx ) * freqV;
@@ -121,7 +121,7 @@ public:
         return DoRemaining( noiseOut, totalValues, index, min, max, gen );
     }
 
-    FastNoise::OutputMinMax GenUniformGrid3D( float* noiseOut, int32_t xStart, int32_t yStart, int32_t zStart, int32_t xSize, int32_t ySize, int32_t zSize, float frequency, int32_t seed ) const final
+    FastNoise::OutputMinMax GenUniformGrid3D( float* noiseOut, int xStart, int yStart, int zStart, int xSize, int ySize, int zSize, float frequency, int seed ) const final
     {
         float32v min( INFINITY );
         float32v max( -INFINITY );
@@ -137,15 +137,15 @@ public:
         int32v ySizeV( ySize );
         int32v yMax = ySizeV + yIdx + int32v( -1 );
 
-        int32_t totalValues = xSize * ySize * zSize;
-        int32_t index = 0;
+        size_t totalValues = xSize * ySize * zSize;
+        size_t index = 0;
 
         xIdx += int32v::FS_Incremented();
 
         AxisReset<true>( xIdx, yIdx, xMax, xSizeV, xSize );
         AxisReset<true>( yIdx, zIdx, yMax, ySizeV, xSize * ySize );
 
-        while( index < totalValues - (int32_t)FS_Size_32() )
+        while( index < totalValues - FS_Size_32() )
         {
             float32v xPos = FS_Converti32_f32( xIdx ) * freqV;
             float32v yPos = FS_Converti32_f32( yIdx ) * freqV;
@@ -175,7 +175,7 @@ public:
         return DoRemaining( noiseOut, totalValues, index, min, max, gen );
     }
 
-    FastNoise::OutputMinMax GenUniformGrid4D( float* noiseOut, int32_t xStart, int32_t yStart, int32_t zStart, int32_t wStart, int32_t xSize, int32_t ySize, int32_t zSize, int32_t wSize, float frequency, int32_t seed ) const final
+    FastNoise::OutputMinMax GenUniformGrid4D( float* noiseOut, int xStart, int yStart, int zStart, int wStart, int xSize, int ySize, int zSize, int wSize, float frequency, int seed ) const final
     {
         float32v min( INFINITY );
         float32v max( -INFINITY );
@@ -194,8 +194,8 @@ public:
         int32v zSizeV( zSize );
         int32v zMax = zSizeV + zIdx + int32v( -1 );
 
-        int32_t totalValues = xSize * ySize * zSize * wSize;
-        int32_t index = 0;
+        size_t totalValues = xSize * ySize * zSize * wSize;
+        size_t index = 0;
 
         xIdx += int32v::FS_Incremented();
 
@@ -203,7 +203,7 @@ public:
         AxisReset<true>( yIdx, zIdx, yMax, ySizeV, xSize * ySize );
         AxisReset<true>( zIdx, wIdx, zMax, zSizeV, xSize * ySize * zSize );
 
-        while( index < totalValues - (int32_t)FS_Size_32() )
+        while( index < totalValues - FS_Size_32() )
         {
             float32v xPos = FS_Converti32_f32( xIdx ) * freqV;
             float32v yPos = FS_Converti32_f32( yIdx ) * freqV;
@@ -236,13 +236,13 @@ public:
         return DoRemaining( noiseOut, totalValues, index, min, max, gen );
     }
 
-    FastNoise::OutputMinMax GenPositionArray2D( float* noiseOut, int32_t count, const float* xPosArray, const float* yPosArray, float xOffset, float yOffset, int32_t seed ) const final
+    FastNoise::OutputMinMax GenPositionArray2D( float* noiseOut, int count, const float* xPosArray, const float* yPosArray, float xOffset, float yOffset, int seed ) const final
     {
         float32v min( INFINITY );
         float32v max( -INFINITY );
 
-        int32_t index = 0;
-        while( index < count - (int32_t)FS_Size_32() )
+        size_t index = 0;
+        while( index < count - FS_Size_32() )
         {
             float32v xPos = float32v( xOffset ) + FS_Load_f32( &xPosArray[index] );
             float32v yPos = float32v( yOffset ) + FS_Load_f32( &yPosArray[index] );
@@ -265,13 +265,13 @@ public:
         return DoRemaining( noiseOut, count, index, min, max, gen );
     }
 
-    FastNoise::OutputMinMax GenPositionArray3D( float* noiseOut, int32_t count, const float* xPosArray, const float* yPosArray, const float* zPosArray, float xOffset, float yOffset, float zOffset, int32_t seed ) const final
+    FastNoise::OutputMinMax GenPositionArray3D( float* noiseOut, int count, const float* xPosArray, const float* yPosArray, const float* zPosArray, float xOffset, float yOffset, float zOffset, int seed ) const final
     {
         float32v min( INFINITY );
         float32v max( -INFINITY );
 
-        int32_t index = 0;
-        while( index < count - (int32_t)FS_Size_32() )
+        size_t index = 0;
+        while( index < count - FS_Size_32() )
         {
             float32v xPos = float32v( xOffset ) + FS_Load_f32( &xPosArray[index] );
             float32v yPos = float32v( yOffset ) + FS_Load_f32( &yPosArray[index] );
@@ -296,13 +296,13 @@ public:
         return DoRemaining( noiseOut, count, index, min, max, gen );
     }
 
-    FastNoise::OutputMinMax GenPositionArray4D( float* noiseOut, int32_t count, const float* xPosArray, const float* yPosArray, const float* zPosArray, const float* wPosArray, float xOffset, float yOffset, float zOffset, float wOffset, int32_t seed ) const final
+    FastNoise::OutputMinMax GenPositionArray4D( float* noiseOut, int count, const float* xPosArray, const float* yPosArray, const float* zPosArray, const float* wPosArray, float xOffset, float yOffset, float zOffset, float wOffset, int seed ) const final
     {
         float32v min( INFINITY );
         float32v max( -INFINITY );
 
-        int32_t index = 0;
-        while( index < count - (int32_t)FS_Size_32() )
+        size_t index = 0;
+        while( index < count - FS_Size_32() )
         {
             float32v xPos = float32v( xOffset ) + FS_Load_f32( &xPosArray[index] );
             float32v yPos = float32v( yOffset ) + FS_Load_f32( &yPosArray[index] );
@@ -329,22 +329,22 @@ public:
         return DoRemaining( noiseOut, count, index, min, max, gen );
     }
 
-    float GenSingle2D( float x, float y, int32_t seed ) const final
+    float GenSingle2D( float x, float y, int seed ) const final
     {
         return FS_Extract0_f32( Gen( int32v( seed ), float32v( x ), float32v( y ) ) );
     }
 
-    float GenSingle3D( float x, float y, float z, int32_t seed ) const final
+    float GenSingle3D( float x, float y, float z, int seed ) const final
     {
         return FS_Extract0_f32( Gen( int32v( seed ), float32v( x ), float32v( y ), float32v( z ) ) );
     }
 
-    float GenSingle4D( float x, float y, float z, float w, int32_t seed ) const final
+    float GenSingle4D( float x, float y, float z, float w, int seed ) const final
     {
         return FS_Extract0_f32( Gen( int32v( seed ), float32v( x ), float32v( y ), float32v( z ), float32v( w ) ) );
     }
 
-    FastNoise::OutputMinMax GenTileable2D( float* noiseOut, int32_t xSize, int32_t ySize, float frequency, int32_t seed ) const final
+    FastNoise::OutputMinMax GenTileable2D( float* noiseOut, int xSize, int ySize, float frequency, int seed ) const final
     {
         float32v min( INFINITY );
         float32v max( -INFINITY );
@@ -356,8 +356,8 @@ public:
         int32v ySizeV( ySize );
         int32v xMax = xSizeV + xIdx + int32v( -1 );
 
-        int32_t totalValues = xSize * ySize;
-        int32_t index = 0;
+        size_t totalValues = xSize * ySize;
+        size_t index = 0;
 
         float pi2Recip( 0.15915493667f );
         float xSizePi = (float)xSize * pi2Recip;
@@ -371,7 +371,7 @@ public:
 
         AxisReset<true>( xIdx, yIdx, xMax, xSizeV, xSize );
 
-        while( index < totalValues - (int32_t)FS_Size_32() )
+        while( index < totalValues - FS_Size_32() )
         {
             float32v xF = FS_Converti32_f32( xIdx ) * xMul;
             float32v yF = FS_Converti32_f32( yIdx ) * yMul;
@@ -410,9 +410,9 @@ public:
 
 private:
     template<bool INITIAL>
-    static FS_INLINE void AxisReset( int32v& aIdx, int32v& bIdx, int32v aMax, int32v aSize, int32_t aStep )
+    static FS_INLINE void AxisReset( int32v& aIdx, int32v& bIdx, int32v aMax, int32v aSize, size_t aStep )
     {
-        for( int32_t resetLoop = INITIAL ? aStep : 0; resetLoop < (int32_t)FS_Size_32(); resetLoop += aStep )
+        for( size_t resetLoop = INITIAL ? aStep : 0; resetLoop < FS_Size_32(); resetLoop += aStep )
         {
             mask32v aReset = aIdx > aMax;
             bIdx = FS_MaskedIncrement_i32( bIdx, aReset );
@@ -420,10 +420,10 @@ private:
         }
     }
 
-    static FS_INLINE FastNoise::OutputMinMax DoRemaining( float* noiseOut, int32_t totalValues, int32_t index, float32v min, float32v max, float32v finalGen )
+    static FS_INLINE FastNoise::OutputMinMax DoRemaining( float* noiseOut, size_t totalValues, size_t index, float32v min, float32v max, float32v finalGen )
     {
         FastNoise::OutputMinMax minMax;
-        int32_t remaining = totalValues - index;
+        size_t remaining = totalValues - index;
 
         if( remaining == FS_Size_32() )
         {
@@ -436,7 +436,7 @@ private:
         }
         else
         {
-            std::memcpy( &noiseOut[index], &finalGen, remaining * sizeof( int32_t ) );
+            std::memcpy( &noiseOut[index], &finalGen, remaining * sizeof( float ) );
 
 #if FASTNOISE_CALC_MIN_MAX
             do
